31. Event System!
I'm overhauling all of the communication between different parts of the
game with a single Game Logic static class (GLogic). This will manager
all of the damage, score, missile count, and other numbers, along with
transitioning between scenes.

Still needs a lot more work, but I'm getting the hang of it.